Move KeyHandle to handle module
