+++ src/asymmetric.rs
@@ -7,10 +7,10 @@
 //!
 //! > **NOTE**: This is currently a stub and should be expanded.
 
-use crate::handle::AlgoHandle;
+use crate::handle::{AlgoHandle, KeyHandle};
 use crate::helpers::blob::{Blob, BlobLayout};
 use crate::helpers::WideCString;
-use crate::key::{BlobType, ErasedKeyBlob, KeyHandle};
+use crate::key::{BlobType, ErasedKeyBlob};
 use crate::property::{Access, AlgorithmName, EccCurveName};
 use crate::Result;
 use std::borrow::Borrow;
