+++ src/asymmetric/builder.rs
@@ -1,9 +1,9 @@
 //! Type-safe builders to generate various asymmetric keys.
 
-use crate::handle::Handle;
+use crate::handle::{Handle, KeyHandle};
 use crate::helpers::{Blob, WideCString};
+use crate::key::BlobType;
 use crate::key::ErasedKeyBlob;
-use crate::key::{BlobType, KeyHandle};
 use crate::{Error, Result};
 use std::marker::PhantomData;
 use std::ptr::null_mut;
