+++ src/handle.rs
@@ -13,6 +13,7 @@ pub trait Handle {
     fn as_mut_ptr(&mut self) -> *mut BCRYPT_HANDLE;
 }
 
+/// Algorithm provider handle
 pub struct AlgoHandle {
     handle: BCRYPT_ALG_HANDLE,
 }
@@ -49,6 +50,45 @@ impl Handle for AlgoHandle {
     }
 
     fn as_mut_ptr(&mut self) -> *mut BCRYPT_ALG_HANDLE {
+        &mut self.handle
+    }
+}
+
+/// Cryptographic key handle
+pub struct KeyHandle {
+    pub(crate) handle: BCRYPT_KEY_HANDLE,
+}
+
+impl KeyHandle {
+    pub fn new() -> Self {
+        Self {
+            handle: ptr::null_mut(),
+        }
+    }
+}
+
+impl Drop for KeyHandle {
+    fn drop(&mut self) {
+        if !self.handle.is_null() {
+            unsafe {
+                BCryptDestroyKey(self.handle);
+            }
+        }
+    }
+}
+
+impl Default for KeyHandle {
+    fn default() -> Self {
+        KeyHandle::new()
+    }
+}
+
+impl Handle for KeyHandle {
+    fn as_ptr(&self) -> BCRYPT_KEY_HANDLE {
+        self.handle
+    }
+
+    fn as_mut_ptr(&mut self) -> *mut BCRYPT_KEY_HANDLE {
         &mut self.handle
     }
 }
