+++ src/symmetric.rs
@@ -43,9 +43,8 @@
 //! [`SymmetricAlgorithm.valid_key_sizes`]: struct.SymmetricAlgorithm.html#method.valid_key_sizes
 
 use crate::buffer::Buffer;
-use crate::handle::{AlgoHandle, Handle};
+use crate::handle::{AlgoHandle, Handle, KeyHandle};
 use crate::helpers::WideCString;
-use crate::key::KeyHandle;
 use crate::property::{self, Access, BlockLength, KeyLength, KeyLengths, ObjectLength};
 use crate::{Error, Result};
 use std::mem::MaybeUninit;
